Reshuffle loading for simpler file loading
