+++ src/ducks/modules/session.js
@@ -1,4 +1,3 @@
-import { actionTypes as protocolActionTypes } from './protocol';
 import { actionTypes as protocolFileActionTypes } from './protocol/file';
 import { actionTypes as protocolStageActionTypes } from './protocol/stages';
 
@@ -10,7 +9,7 @@ const initialState = {
 
 export default function reducer(state = initialState, action = {}) {
   switch (action.type) {
-    case protocolActionTypes.SET_PROTOCOL: {
+    case protocolFileActionTypes.OPEN_PROTOCOL: {
       return {
         ...state,
         activeProtocol: action.meta,
@@ -18,16 +17,16 @@ export default function reducer(state = initialState, action = {}) {
         lastChanged: 0,
       };
     }
+    case protocolFileActionTypes.SAVE_COMPLETE:
+      return {
+        ...state,
+        lastSaved: new Date().getTime(),
+      };
     case protocolStageActionTypes.CREATE_STAGE:
     case protocolStageActionTypes.UPDATE_STAGE:
       return {
         ...state,
         lastChanged: new Date().getTime(),
-      };
-    case protocolFileActionTypes.SAVE_COMPLETE:
-      return {
-        ...state,
-        lastSaved: new Date().getTime(),
       };
     default:
       return state;
