+++ src/history.js
@@ -1,0 +1,5 @@
+import createHistory from 'history/createMemoryHistory';
+
+const history = createHistory();
+
+export default history;
