+++ src/index.js
@@ -3,8 +3,9 @@ import ReactDOM from 'react-dom';
 import injectTapEventPlugin from 'react-tap-event-plugin';
 import initReactFastclick from 'react-fastclick';
 import { Provider } from 'react-redux';
-import { MemoryRouter as Router, Route } from 'react-router-dom';
+import { Router, Route } from 'react-router-dom';
 import './styles/main.scss';
+import memoryHistory from './history';
 import { store } from './ducks/store';
 import App from './components/App';
 import Routes from './routes';
@@ -18,7 +19,7 @@ const startApp = () => {
     [
       <ClipPaths />,
       <Provider store={store}>
-        <Router>
+        <Router history={memoryHistory}>
           <Route
             render={({ location, history }) => (
               <App>
